requirements.txt
streamlit==1.42.0
pandas==2.2.3
numpy==2.2.3
openpyxl==3.1.5
watchdog==6.0.0
google-api-python-client==
google-auth-httplib2==
google-auth-oauthlib==
